cleanup RouteBundleExtras
